requirements.txt
appdata==2.1.2
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.6
colored==1.4.3
customtkinter==5.0.3
darkdetect==0.8.0
docopt==0.6.2
idna==3.4
numpy==1.23.4
pandas==1.5.1
Pillow==9.4.0
pipreqs==0.4.11
pyperclip==1.8.2
pyreadline3==3.4.1
python-dateutil==2.8.2
pytube==12.1.2
pytz==2022.6
requests==2.28.1
requestsexceptions==1.4.0
six==1.16.0
soupsieve==2.3.2.post1
to-requirements.txt==1.3.0
urllib3==1.26.12
wcwidth==0.2.5
yarg==0.1.9
